requirements.txt
flask
python-dotenv
requests
gunicorn
pillow
openai
pytesseract